Move initialise Reporting tables and indicators to ChwRepository class
