+++ opensrp-chw/build.gradle
@@ -149,7 +149,7 @@ dependencies {
         exclude group: 'com.android.support', module: 'appcompat-v7'
     }
 
-    implementation('org.smartregister:opensrp-client-reporting:1.0.1-SNAPSHOT@aar') {
+    implementation('org.smartregister:opensrp-client-reporting:1.0.5-SNAPSHOT@aar') {
         transitive = true
         exclude group: 'org.smartregister', module: 'opensrp-client-core'
         exclude group: 'com.android.support', module: 'appcompat-v7'
