+++ opensrp-chw/src/main/java/org/smartregister/chw/application/ChwApplication.java
@@ -66,10 +66,6 @@ public class ChwApplication extends DrishtiApplication {
 
     private RulesEngineHelper rulesEngineHelper;
 
-    private String indicatorsConfigFile = "config/indicator-definitions.yml";
-    private String indicatorDataInitialisedPref = "INDICATOR_DATA_INITIALISED";
-    private String appVersionCodePref = "APP_VERSION_CODE";
-
     public static synchronized ChwApplication getInstance() {
         return (ChwApplication) mInstance;
     }
@@ -146,16 +142,6 @@ public class ChwApplication extends DrishtiApplication {
 
         // Init Reporting library
         ReportingLibrary.init(context, getRepository(), null, BuildConfig.VERSION_CODE, BuildConfig.DATABASE_VERSION);
-        ReportingLibrary reportingLibraryInstance = ReportingLibrary.getInstance();
-        // Check if indicator data initialised
-        boolean indicatorDataInitialised = Boolean.parseBoolean(reportingLibraryInstance.getContext()
-                .allSharedPreferences().getPreference(indicatorDataInitialisedPref));
-        boolean isUpdated = checkIfAppUpdated();
-        if (!indicatorDataInitialised || isUpdated) {
-            reportingLibraryInstance.initIndicatorData(indicatorsConfigFile); // This will persist the data in the DB
-            reportingLibraryInstance.getContext().allSharedPreferences().savePreference(indicatorDataInitialisedPref, "true");
-            reportingLibraryInstance.getContext().allSharedPreferences().savePreference(appVersionCodePref, String.valueOf(BuildConfig.VERSION_CODE));
-        }
 
         //init Job Manager
         JobManager.create(this).addJobCreator(new ChwJobCreator());
@@ -176,16 +162,6 @@ public class ChwApplication extends DrishtiApplication {
         }
         if (language.equals(Locale.FRENCH.getLanguage()))
             saveLanguage(Locale.FRENCH.getLanguage());
-    }
-
-    private boolean checkIfAppUpdated() {
-        String savedAppVersion = ReportingLibrary.getInstance().getContext().allSharedPreferences().getPreference(appVersionCodePref);
-        if (savedAppVersion.isEmpty()) {
-            return true;
-        } else {
-            int savedVersion = Integer.parseInt(savedAppVersion);
-            return (BuildConfig.VERSION_CODE > savedVersion);
-        }
     }
 
     private void saveLanguage(String language) {
