+++ opensrp-chw/src/main/java/org/smartregister/chw/repository/ChwRepository.java
@@ -20,6 +20,10 @@ import org.smartregister.immunization.repository.VaccineNameRepository;
 import org.smartregister.immunization.repository.VaccineRepository;
 import org.smartregister.immunization.repository.VaccineTypeRepository;
 import org.smartregister.immunization.util.IMDatabaseUtils;
+import org.smartregister.reporting.ReportingLibrary;
+import org.smartregister.reporting.repository.DailyIndicatorCountRepository;
+import org.smartregister.reporting.repository.IndicatorQueryRepository;
+import org.smartregister.reporting.repository.IndicatorRepository;
 import org.smartregister.repository.AlertRepository;
 import org.smartregister.repository.EventClientRepository;
 import org.smartregister.repository.Repository;
@@ -35,6 +39,9 @@ public class ChwRepository extends Repository {
     protected SQLiteDatabase readableDatabase;
     protected SQLiteDatabase writableDatabase;
     private Context context;
+    private String indicatorsConfigFile = "config/indicator-definitions.yml";
+    private String indicatorDataInitialisedPref = "INDICATOR_DATA_INITIALISED";
+    private String appVersionCodePref = "APP_VERSION_CODE";
 
     public ChwRepository(Context context, org.smartregister.Context openSRPContext) {
         super(context, AllConstants.DATABASE_NAME, BuildConfig.DATABASE_VERSION, openSRPContext.session(), ChwApplication.createCommonFtsObject(), openSRPContext.sharedRepositoriesArray());
@@ -57,12 +64,27 @@ public class ChwRepository extends Repository {
 
         RecurringServiceTypeRepository.createTable(database);
         RecurringServiceRecordRepository.createTable(database);
+
+        IndicatorRepository.createTable(database);
+        IndicatorQueryRepository.createTable(database);
+        DailyIndicatorCountRepository.createTable(database);
+
         //onUpgrade(database, 1, 2);
         RecurringServiceTypeRepository recurringServiceTypeRepository = ImmunizationLibrary.getInstance().recurringServiceTypeRepository();
         IMDatabaseUtils.populateRecurringServices(context, database, recurringServiceTypeRepository);
 
-
         onUpgrade(database, 1, BuildConfig.DATABASE_VERSION);
+
+        ReportingLibrary reportingLibraryInstance = ReportingLibrary.getInstance();
+        // Check if indicator data initialised
+        boolean indicatorDataInitialised = Boolean.parseBoolean(reportingLibraryInstance.getContext()
+                .allSharedPreferences().getPreference(indicatorDataInitialisedPref));
+        boolean isUpdated = checkIfAppUpdated();
+        if (!indicatorDataInitialised || isUpdated) {
+            reportingLibraryInstance.initIndicatorData(indicatorsConfigFile, database); // This will persist the data in the DB
+            reportingLibraryInstance.getContext().allSharedPreferences().savePreference(indicatorDataInitialisedPref, "true");
+            reportingLibraryInstance.getContext().allSharedPreferences().savePreference(appVersionCodePref, String.valueOf(BuildConfig.VERSION_CODE));
+        }
 
     }
 
@@ -103,9 +125,9 @@ public class ChwRepository extends Repository {
     @Override
     public SQLiteDatabase getReadableDatabase() {
         String pass = ChwApplication.getInstance().getPassword();
-        if(StringUtils.isNotBlank(pass)){
+        if (StringUtils.isNotBlank(pass)) {
             return getReadableDatabase(pass);
-        }else{
+        } else {
             throw new IllegalStateException("Password is blank");
         }
     }
@@ -113,9 +135,9 @@ public class ChwRepository extends Repository {
     @Override
     public SQLiteDatabase getWritableDatabase() {
         String pass = ChwApplication.getInstance().getPassword();
-        if(StringUtils.isNotBlank(pass)){
+        if (StringUtils.isNotBlank(pass)) {
             return getWritableDatabase(pass);
-        }else{
+        } else {
             throw new IllegalStateException("Password is blank");
         }
     }
@@ -158,6 +180,16 @@ public class ChwRepository extends Repository {
             writableDatabase.close();
         }
         super.close();
+    }
+
+    private boolean checkIfAppUpdated() {
+        String savedAppVersion = ReportingLibrary.getInstance().getContext().allSharedPreferences().getPreference(appVersionCodePref);
+        if (savedAppVersion.isEmpty()) {
+            return true;
+        } else {
+            int savedVersion = Integer.parseInt(savedAppVersion);
+            return (BuildConfig.VERSION_CODE > savedVersion);
+        }
     }
 
     private void upgradeToVersion2(SQLiteDatabase db) {
@@ -229,8 +261,8 @@ public class ChwRepository extends Repository {
 
     private void upgradeToVersion6(SQLiteDatabase db) {
         try {
-            if(BuildConfig.BUILD_COUNTRY == Country.TANZANIA){
-                for (String query: RepositoryUtils.UPGRADE_V6) {
+            if (BuildConfig.BUILD_COUNTRY == Country.TANZANIA) {
+                for (String query : RepositoryUtils.UPGRADE_V6) {
                     db.execSQL(query);
                 }
             }
@@ -238,6 +270,7 @@ public class ChwRepository extends Repository {
             Log.e(TAG, "upgradeToVersion6 " + Log.getStackTraceString(e));
         }
     }
+
     private void upgradeToVersion7(SQLiteDatabase db) {
         try {
             db.execSQL(HomeVisitRepository.UPDATE_TABLE_ADD_VACCINE_NOT_GIVEN);
